+++ src/VariableSizeList.js
@@ -205,6 +205,17 @@ const VariableSizeList = createListComponent({
       itemMetadata.offset - size + itemMetadata.size
     );
 
+    if (align === 'smart') {
+      if (
+        scrollOffset >= minOffset - size &&
+        scrollOffset <= maxOffset + size
+      ) {
+        align = 'auto';
+      } else {
+        align = 'center';
+      }
+    }
+
     switch (align) {
       case 'start':
         return maxOffset;