+++ src/__tests__/FixedSizeGrid.js
@@ -27,6 +27,31 @@ describe('FixedSizeGrid', () => {
   beforeEach(() => {
     jest.useFakeTimers();
 
+    // JSdom does not do actual layout and so doesn't return meaningful values here.
+    // For the purposes of our tests though, we can mock out semi-meaningful values.
+    Object.defineProperties(HTMLElement.prototype, {
+      clientWidth: {
+        configurable: true,
+        get: function() {
+          return parseInt(this.style.width, 10) || 0;
+        },
+      },
+      clientHeight: {
+        configurable: true,
+        get: function() {
+          return parseInt(this.style.height, 10) || 0;
+        },
+      },
+      scrollHeight: {
+        configurable: true,
+        get: () => Number.MAX_SAFE_INTEGER,
+      },
+      scrollWidth: {
+        configurable: true,
+        get: () => Number.MAX_SAFE_INTEGER,
+      },
+    });
+
     // Mock the DOM helper util for testing purposes.
     getScrollbarSize = domHelpers.getScrollbarSize = jest.fn(() => 0);
 
@@ -99,6 +124,7 @@ describe('FixedSizeGrid', () => {
       // Scroll, then capture the rendered style for item 1,
       // Then let the debounce timer clear the cached styles.
       simulateScroll(instance, { scrollLeft: 100, scrollTop: 25 });
+      expect(itemRenderer).toHaveBeenCalled();
       const itemOneArgsA = itemRenderer.mock.calls.find(
         ([params]) => params.columnIndex === 1 && params.rowIndex === 1
       );
@@ -107,6 +133,7 @@ describe('FixedSizeGrid', () => {
       // Scroll again, then capture the rendered style for item 1,
       // And confirm that the style was recreated.
       simulateScroll(instance, { scrollLeft: 0, scrollTop: 0 });
+      expect(itemRenderer).toHaveBeenCalled();
       const itemOneArgsB = itemRenderer.mock.calls.find(
         ([params]) => params.columnIndex === 1 && params.rowIndex === 1
       );
@@ -155,7 +182,7 @@ describe('FixedSizeGrid', () => {
       <FixedSizeGrid {...defaultProps} />
     );
     const scrollContainer = findScrollContainer(rendered);
-    expect(scrollContainer.props.style.pointerEvents).toBe('');
+    expect(scrollContainer.props.style.pointerEvents).toBe(undefined);
     rendered.getInstance().setState({ isScrolling: true });
     expect(scrollContainer.props.style.pointerEvents).toBe('none');
   });
@@ -211,15 +238,15 @@ describe('FixedSizeGrid', () => {
     });
   });
 
-  describe('overscanColumnsCount and overscanRowsCount', () => {
+  describe('overscanColumnCount and overscanRowCount', () => {
     it('should require a minimum of 1 overscan to support tabbing', () => {
       ReactTestRenderer.create(
         <FixedSizeGrid
           {...defaultProps}
           initialScrollLeft={250}
           initialScrollTop={250}
-          overscanColumnsCount={0}
-          overscanRowsCount={0}
+          overscanColumnCount={0}
+          overscanRowCount={0}
         />
       );
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
@@ -231,8 +258,8 @@ describe('FixedSizeGrid', () => {
           {...defaultProps}
           initialScrollLeft={250}
           initialScrollTop={250}
-          overscanColumnsCount={2}
-          overscanRowsCount={2}
+          overscanColumnCount={2}
+          overscanRowCount={2}
         />
       );
       rendered.getInstance().scrollTo({ scrollLeft: 1000, scrollTop: 1000 });
@@ -246,8 +273,8 @@ describe('FixedSizeGrid', () => {
           {...defaultProps}
           initialScrollLeft={250}
           initialScrollTop={250}
-          overscanColumnsCount={2}
-          overscanRowsCount={2}
+          overscanColumnCount={2}
+          overscanRowCount={2}
         />
       );
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
@@ -259,8 +286,8 @@ describe('FixedSizeGrid', () => {
           {...defaultProps}
           initialScrollLeft={250}
           initialScrollTop={250}
-          overscanColumnsCount={2}
-          overscanRowsCount={2}
+          overscanColumnCount={2}
+          overscanRowCount={2}
         />
       );
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
@@ -294,10 +321,38 @@ describe('FixedSizeGrid', () => {
         expect(console.warn).toHaveBeenCalledTimes(1);
         expect(console.warn).toHaveBeenLastCalledWith(
           'The overscanCount prop has been deprecated. ' +
-            'Please use the overscanColumnsCount and overscanRowsCount props instead.'
+            'Please use the overscanColumnCount and overscanRowCount props instead.'
         );
 
         renderer.update(<FixedSizeGrid {...defaultProps} overscanCount={1} />);
+
+        // But it should only warn once.
+        expect(console.warn).toHaveBeenCalledTimes(1);
+      });
+
+      it('should warn about deprecated overscanRowsCount or overscanColumnsCount prop', () => {
+        spyOn(console, 'warn');
+
+        const renderer = ReactTestRenderer.create(
+          <FixedSizeGrid
+            {...defaultProps}
+            overscanRowsCount={1}
+            overscanColumnsCount={1}
+          />
+        );
+        expect(console.warn).toHaveBeenCalledTimes(1);
+        expect(console.warn).toHaveBeenLastCalledWith(
+          'The overscanColumnsCount and overscanRowsCount props have been deprecated. ' +
+            'Please use the overscanColumnCount and overscanRowCount props instead.'
+        );
+
+        renderer.update(
+          <FixedSizeGrid
+            {...defaultProps}
+            overscanRowsCount={1}
+            overscanColumnsCount={1}
+          />
+        );
 
         // But it should only warn once.
         expect(console.warn).toHaveBeenCalledTimes(1);
@@ -318,7 +373,7 @@ describe('FixedSizeGrid', () => {
         expect(onItemsRendered.mock.calls).toMatchSnapshot();
       });
 
-      it('should use overscanRowsCount if both it and overscanCount are provided', () => {
+      it('should use overscanRowCount if both it and overscanCount are provided', () => {
         spyOn(console, 'warn');
 
         ReactTestRenderer.create(
@@ -327,7 +382,25 @@ describe('FixedSizeGrid', () => {
             initialScrollLeft={100}
             initialScrollTop={100}
             overscanCount={2}
-            overscanRowsCount={3}
+            overscanRowCount={3}
+          />
+        );
+        expect(onItemsRendered.mock.calls).toMatchSnapshot();
+      });
+
+      it('should use overscanColumnCount and overscanRowCount if both them and deprecated props are provided', () => {
+        spyOn(console, 'warn');
+
+        ReactTestRenderer.create(
+          <FixedSizeGrid
+            {...defaultProps}
+            initialScrollLeft={100}
+            initialScrollTop={100}
+            overscanCount={1}
+            overscanColumnsCount={2}
+            overscanColumnCount={3}
+            overscanRowsCount={2}
+            overscanRowCount={3}
           />
         );
         expect(onItemsRendered.mock.calls).toMatchSnapshot();
@@ -342,6 +415,21 @@ describe('FixedSizeGrid', () => {
             initialScrollLeft={100}
             initialScrollTop={100}
             overscanCount={2}
+          />
+        );
+        expect(onItemsRendered.mock.calls).toMatchSnapshot();
+      });
+
+      it('should support deprecated overscanColumnsCount and overscanRowsCount', () => {
+        spyOn(console, 'warn');
+
+        ReactTestRenderer.create(
+          <FixedSizeGrid
+            {...defaultProps}
+            initialScrollLeft={100}
+            initialScrollTop={100}
+            overscanColumnsCount={2}
+            overscanRowsCount={2}
           />
         );
         expect(onItemsRendered.mock.calls).toMatchSnapshot();
@@ -431,31 +519,30 @@ describe('FixedSizeGrid', () => {
         })
       );
     });
+
+    it('should ignore offsets less than zero', () => {
+      const onScroll = jest.fn();
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} onScroll={onScroll} />,
+        document.createElement('div')
+      );
+      instance.scrollTo({ scrollLeft: 100, scrollTop: 100 });
+      onScroll.mockClear();
+      instance.scrollTo({ scrollLeft: -1, scrollTop: -1 });
+      expect(onScroll.mock.calls[0][0].scrollLeft).toBe(0);
+      expect(onScroll.mock.calls[0][0].scrollTop).toBe(0);
+    });
   });
 
   describe('scrollToItem method', () => {
     it('should not set invalid offsets when the list contains few items', () => {
-      const onScroll = jest.fn();
       const rendered = ReactTestRenderer.create(
-        <FixedSizeGrid
-          {...defaultProps}
-          columnCount={1}
-          rowCount={2}
-          onScroll={onScroll}
-        />
+        <FixedSizeGrid {...defaultProps} columnCount={1} rowCount={2} />
       );
-      onScroll.mockClear();
-      // Offset should not be negative.
-      rendered
-        .getInstance()
-        .scrollToItem({ columnIndex: 0, rowIndex: 0, align: 'auto' });
-      expect(onScroll).toHaveBeenCalledWith({
-        horizontalScrollDirection: 'backward',
-        scrollLeft: 0,
-        scrollTop: 0,
-        scrollUpdateWasRequested: true,
-        verticalScrollDirection: 'backward',
-      });
+      expect(onItemsRendered).toMatchSnapshot();
+      onItemsRendered.mockClear();
+      rendered.getInstance().scrollToItem(0);
+      expect(onItemsRendered).not.toHaveBeenCalled();
     });
 
     it('should scroll to the correct item for align = "auto"', () => {
@@ -563,6 +650,41 @@ describe('FixedSizeGrid', () => {
       rendered.getInstance().scrollToItem({ rowIndex: 10, align: 'center' });
       // Scroll left to column 3, without changing scrollTop
       rendered.getInstance().scrollToItem({ columnIndex: 3, align: 'center' });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
+    });
+
+    it('should scroll to the correct item for align = "smart"', () => {
+      const rendered = ReactTestRenderer.create(
+        <FixedSizeGrid {...defaultProps} />
+      );
+
+      // Scroll down enough to show item 10 at the center.
+      // It was further than one screen away, so it gets centered.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 10, rowIndex: 10, align: 'smart' });
+      // No need to scroll again; item 9 is already visible.
+      // Overscan indices will change though, since direction changes.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 9, rowIndex: 9, align: 'smart' });
+      // Scroll up enough to show item 2 as close to the center as we can.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 2, rowIndex: 2, align: 'smart' });
+      // Scroll down to row 10, without changing scrollLeft
+      rendered.getInstance().scrollToItem({ rowIndex: 10, align: 'smart' });
+      // Scroll left to column 0, without changing scrollTop
+      rendered.getInstance().scrollToItem({ columnIndex: 0, align: 'smart' });
+
+      // Scrolling within a distance of a single screen from viewport
+      // should have the 'auto' behavior of scrolling as little as possible.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 5, rowIndex: 5, align: 'smart' });
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 10, rowIndex: 10, align: 'smart' });
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
     });
 
@@ -678,6 +800,30 @@ describe('FixedSizeGrid', () => {
         scrollUpdateWasRequested: true,
         verticalScrollDirection: 'backward',
       });
+    });
+
+    it('should ignore indexes less than zero', () => {
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} />,
+        document.createElement('div')
+      );
+      instance.scrollToItem({ columnIndex: 20, rowIndex: 20 });
+      onItemsRendered.mockClear();
+      instance.scrollToItem({ columnIndex: -1, rowIndex: -1 });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
+    });
+
+    it('should ignore indexes greater than itemCount', () => {
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} />,
+        document.createElement('div')
+      );
+      onItemsRendered.mockClear();
+      instance.scrollToItem({
+        columnIndex: defaultProps.columnCount * 2,
+        rowIndex: defaultProps.rowCount * 2,
+      });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
     });
   });