+++ website/src/routes/api/FixedSizeList.js
@@ -403,6 +403,12 @@ const METHODS = [
             auto (default) - Scroll as little as possible to ensure the item is
             visible. (If the item is already visible, it won't scroll at all.)
           </li>
+          <li>
+            smart - If the item is already visible, don't scroll at all. If it
+            is less than one viewport away, scroll as little as possible so that
+            it becomes visible. If it is more than one viewport away, scroll so
+            that it is centered within the list.
+          </li>
           <li>center - Center align the item within the list.</li>
           <li>
             end - Align the item to the end of the list (the bottom for vertical